+++ src/main/java/com/solosw/codelab/controller/PublicKeyController.java
@@ -7,6 +7,7 @@ import com.solosw.codelab.entity.po.Users;
 import com.solosw.codelab.service.PublicKeyService;
 import com.solosw.codelab.service.UsersService;
 import com.solosw.codelab.utils.GitoliteUtil;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -26,8 +27,11 @@ public class PublicKeyController extends BaseController {
     UsersService usersService;
 
     @PostMapping("/createPublicKey")
-    public ResponseBo createPublicKey(@RequestBody Map<String,Object> map) throws Exception {
+    public ResponseBo createPublicKey(@RequestBody Map<String,Object> map, HttpServletRequest request) throws Exception {
+
         Integer creator=(Integer) map.get("creatorId");
+        if(getCurrentUser(request).getId().intValue()!=(creator))return ResponseBo.getFail(null,"错误操作",500);
+
         String publicKey=(String) map.get("publicKey");
         String result = publicKey.replaceAll("[\\s\\n\\r]", "");
         String keyName= (String) map.get("keyName");
@@ -37,13 +41,17 @@ public class PublicKeyController extends BaseController {
     }
 
     @PostMapping("/getUserKey/{creatorId}")
-    public ResponseBo getUserKey(@PathVariable Long creatorId){
+    public ResponseBo getUserKey(@PathVariable Long creatorId, HttpServletRequest request){
+        if(!getCurrentUser(request).getId().equals(creatorId))return ResponseBo.getFail(null,"错误操作",500);
+
         List<PublicKey> p=publicKeyService.getPublicKeyByCreatorIdOrId(creatorId,0);
         return p==null?ResponseBo.getSuccess(new PublicKey()):ResponseBo.getSuccess(p);
     }
 
     @PostMapping("/deleteKey/{creatorId}/{keyId}")
-    public ResponseBo deleteKey(@PathVariable Long creatorId,@PathVariable Long keyId){
+    public ResponseBo deleteKey(@PathVariable Long creatorId,@PathVariable Long keyId, HttpServletRequest request){
+        if(!getCurrentUser(request).getId().equals(creatorId))return ResponseBo.getFail(null,"错误操作",500);
+
         publicKeyService.deleteById(keyId);
         return ResponseBo.getSuccess(null);
     }
